video_story_generator/requirements.txt
PyQt6>=6.6.0
yt-dlp>=2023.12.0
opencv-python>=4.8.0
scenedetect>=0.6.2
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0
ffmpeg-python>=0.2.0
pydub>=0.25.1
requests>=2.31.0
numpy>=1.24.0
Pillow>=10.0.0




